app/core/consistency.py
from typing import Dict, Any, Optional
import asyncio
from datetime import datetime
from redis import Redis
from app.core.config import settings
from app.core.logging import logger

class DataConsistencyManager:
    def __init__(self):
        self.redis = Redis(
            host=settings.REDIS_HOST,
            port=settings.REDIS_PORT,
            db=settings.REDIS_DB,
            decode_responses=True
        )
        self.locks: Dict[str, Dict] = {}
        self.version_control: Dict[str, int] = {}
        self.pending_changes: Dict[str, List[Dict]] = {}

    async def acquire_lock(self, resource_id: str, user_id: str, timeout: int = 30) -> bool:
        """Optimistic locking için kilit al"""
        try:
            lock_key = f"lock:{resource_id}"
            if await self.redis.setnx(lock_key, user_id):
                await self.redis.expire(lock_key, timeout)
                return True
            return False
        except Exception as e:
            logger.error(f"Error acquiring lock: {str(e)}")
            return False

    async def release_lock(self, resource_id: str, user_id: str) -> bool:
        """Kilidi serbest bırak"""
        try:
            lock_key = f"lock:{resource_id}"
            current_owner = await self.redis.get(lock_key)
            if current_owner == user_id:
                await self.redis.delete(lock_key)
                return True
            return False
        except Exception as e:
            logger.error(f"Error releasing lock: {str(e)}")
            return False

    async def check_version(self, resource_id: str, version: int) -> bool:
        """Versiyon kontrolü yap"""
        try:
            current_version = self.version_control.get(resource_id, 0)
            return version == current_version
        except Exception as e:
            logger.error(f"Error checking version: {str(e)}")
            return False

    async def update_version(self, resource_id: str) -> int:
        """Versiyonu güncelle"""
        try:
            current_version = self.version_control.get(resource_id, 0)
            new_version = current_version + 1
            self.version_control[resource_id] = new_version
            return new_version
        except Exception as e:
            logger.error(f"Error updating version: {str(e)}")
            return 0

    async def queue_change(self, resource_id: str, change: Dict[str, Any]):
        """Değişikliği kuyruğa al"""
        try:
            if resource_id not in self.pending_changes:
                self.pending_changes[resource_id] = []
            self.pending_changes[resource_id].append({
                "change": change,
                "timestamp": datetime.now().isoformat(),
                "version": self.version_control.get(resource_id, 0)
            })
        except Exception as e:
            logger.error(f"Error queueing change: {str(e)}")

    async def apply_changes(self, resource_id: str) -> bool:
        """Kuyruktaki değişiklikleri uygula"""
        try:
            if resource_id not in self.pending_changes:
                return True

            changes = self.pending_changes[resource_id]
            for change in changes:
                if await self.check_version(resource_id, change["version"]):
                    # Değişikliği uygula
                    await self._apply_single_change(resource_id, change["change"])
                    await self.update_version(resource_id)
                else:
                    # Versiyon uyuşmazlığı, çakışma çözümle
                    await self._resolve_conflict(resource_id, change)

            self.pending_changes[resource_id] = []
            return True
        except Exception as e:
            logger.error(f"Error applying changes: {str(e)}")
            return False

    async def _apply_single_change(self, resource_id: str, change: Dict[str, Any]):
        """Tek bir değişikliği uygula"""
        # Değişikliği uygula
        pass

    async def _resolve_conflict(self, resource_id: str, change: Dict[str, Any]):
        """Çakışmayı çöz"""
        # Çakışma çözümleme stratejisi uygula
        pass

    async def rollback_changes(self, resource_id: str):
        """Değişiklikleri geri al"""
        try:
            if resource_id in self.pending_changes:
                self.pending_changes[resource_id] = []
            # Son başarılı duruma geri dön
            pass
        except Exception as e:
            logger.error(f"Error rolling back changes: {str(e)}")

    def get_resource_status(self, resource_id: str) -> Dict:
        """Kaynak durumunu raporla"""
        return {
            "version": self.version_control.get(resource_id, 0),
            "locked": resource_id in self.locks,
            "pending_changes": len(self.pending_changes.get(resource_id, [])),
            "last_update": datetime.now().isoformat()
        }

consistency_manager = DataConsistencyManager() 